fix: refactor nocobase test
