+++ packages/core/client/src/appInfo/__tests__/CurrentAppInfoProvider.test.tsx
@@ -3,13 +3,17 @@ import { CurrentAppInfoProvider, useCurrentAppInfo } from '@nocobase/client';
 
 describe('CurrentAppInfoProvider', () => {
   it('should work', async () => {
-    const { result } = await renderHookWithApp(() => useCurrentAppInfo(), CurrentAppInfoProvider, {
-      'app:getInfo': {
-        database: {
-          dialect: 'mysql',
+    const { result } = await renderHookWithApp({
+      hook: useCurrentAppInfo,
+      Wrapper: CurrentAppInfoProvider,
+      apis: {
+        'app:getInfo': {
+          database: {
+            dialect: 'mysql',
+          },
+          lang: 'zh-CN',
+          version: '1.0.0',
         },
-        lang: 'zh-CN',
-        version: '1.0.0',
       },
     });
     await waitFor(() => {
