+++ packages/core/client/src/schema-component/antd/pagination/__tests__/pagination.test.tsx
@@ -1,9 +1,9 @@
-import { renderSchema } from '@nocobase/test/client';
+import { renderApp } from '@nocobase/test/client';
 import { Pagination } from '@nocobase/client';
 
 describe('Pagination', () => {
   it('renders without errors', async () => {
-    const { container } = await renderSchema({
+    const { container } = await renderApp({
       Component: Pagination,
       props: {
         total: 20,
@@ -113,7 +113,7 @@ describe('Pagination', () => {
   });
 
   it('hides when hidden prop is true', async () => {
-    const { container } = await renderSchema({
+    const { container } = await renderApp({
       Component: Pagination,
       props: {
         hidden: true,
