+++ packages/core/test/src/web/dataSource2.json
