+++ packages/core/test/src/web/dataSourceMainCollections.json
