+++ packages/core/test/web.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/web';
