+++ packages/core/test/web.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/web');
